Removed unneccesary parts - corrected scaling issues
Removed the no-PCB variants of the parts. Separated out the PCB mounting holes to allow it to be used / moved independently of the electrical connectors included in the servo parts.
Also corrected some scaling issues caused by inkscape conversion.
Enhanced the connector icons and added descriptions.
Revised the file naming conventions for consistency.