Added JST-VH 3 pin male header and female connector types
These connectors are commonly used as connectors for Dagu robot main
motors. They do not play nicely with 0.1 in strip or breadboard. The male
can be mounted on stripboard if the centre hole is used for location purposes,
holes drilled between the tracks and solder traces added to make high
current double tracks for the outer connectors.

The Fritzing part in this case has pin0 as the centre pin so that the male
header locates symmetrically on stripboard. The labelling is still pin 1
is the leftmost however.